refactor: extract util.rs from lib.rs (#112 PR 2)
Move shared utility functions from lib.rs to src/util.rs:

- usage_text, print_usage: CLI usage text
- parse_config_flag: --config argument parsing
- binary_name, clone_lossy: OsString helpers
- should_block_for_sudo + libc_geteuid: platform-specific sudo detection
- resolve_real_command, resolve_real_command_from_path: PATH resolution

3 tests moved from lib.rs to util.rs (binary_name, sudo_block,
resolve_real_command). Total test count unchanged at 473.

lib.rs: -154 lines (2,534 → ~2,380)
